# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: AstraCalls API
  version: "1.0.2"
  description: |
    API HTTP do **AstraCalls** — chamadas de voz do WhatsApp pelo navegador.

    ## Autenticação
    Não há login/senha. As rotas `/api/*` são protegidas por API key: envie o header
    **`X-API-Key`** com o valor definido em `WACALLS_API_KEY` (ou `?apiKey=...` na query).
    Se `WACALLS_API_KEY` estiver vazia, a autenticação fica desligada.

    O header **`X-Client-Id`** é opcional e **não autentica nada** — é só um identificador
    estável usado para marcar o "dono" das chamadas e filtrar eventos.

    > ⚠️ A API inicia/controla as chamadas (sinalização), mas o **áudio** exige um cliente
    > WebRTC (navegador) com captura de microfone — não é possível só com chamadas HTTP.
servers:
  - url: /
    description: Mesma origem do painel
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key definida em WACALLS_API_KEY (vazia = sem autenticação)"
  parameters:
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema: { type: string, example: meuapp }
      description: >-
        Opcional. Identificador do cliente (dono da chamada / filtro de eventos).
        NÃO é autenticação — para autenticar use o header X-API-Key.
    Sid:
      name: sid
      in: path
      required: true
      schema: { type: string }
      description: ID da sessão (conta WhatsApp).
...
  responses:
    SendOK:
      description: Mensagem enviada
      content:
        application/json:
          schema:
            type: object
            properties:
              id: { type: string, description: "ID da mensagem no WhatsApp" }
              to: { type: string }
              timestamp: { type: integer, format: int64 }
...
paths:
...
  /api/sessions/{sid}/messages/form:
    post:
      tags: [Interativas]
      summary: Formulário (webview) — abre uma página de form DENTRO do WhatsApp
      description: >
        Envia uma mensagem interativa com um botão que abre uma WEBVIEW em tela
        cheia dentro do próprio WhatsApp (não é o navegador externo), renderizando
        um formulário com os campos informados. Ao enviar, as respostas: (1) caem
        no Chatwoot como mensagem recebida do cliente, (2) disparam o webhook
        `form_response` com os dados estruturados. Alternativa ao WhatsApp Flows
        nativo (que exige conta oficial/WABA e NÃO funciona em número não-oficial).
        As rotas `GET /forms/{token}` e `POST /forms/{token}/submit` são públicas
        e não exigem API key. O token é uma capability aleatória, opaca, de uso
        único e armazenada somente como hash. O formulário expira em 24 horas por
        padrão, com validade configurável entre 5 minutos e 7 dias. A resposta de
        envio inclui `form_id` e `expires_at` para correlação com o webhook.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, fields]
              properties:
                to: { type: string, example: "5561999999999" }
                title: { type: string, example: "Cadastro rápido" }
                body: { type: string, example: "Preencha seus dados 👇" }
                footer: { type: string, example: "AstraCalls" }
                buttonText: { type: string, example: "Abrir formulário", description: "Texto do botão que abre a webview" }
                submitText: { type: string, example: "Enviar cadastro", description: "Texto do botão de envio dentro do form" }
                expiresInSeconds: { type: integer, minimum: 300, maximum: 604800, default: 86400, description: "Validade do link em segundos" }
                fields:
                  type: array
                  description: "Campos do formulário, na ordem em que aparecem."
                  items:
                    type: object
                    required: [name]
                    properties:
                      name: { type: string, example: "email", description: "Chave do campo (vem no webhook)" }
                      label: { type: string, example: "E-mail" }
                      type:
                        type: string
                        enum: [text, email, tel, number, date, textarea, select]
                        default: text
                      options:
                        type: array
                        items: { type: string }
                        description: "Opções (apenas para type=select)"
                      placeholder: { type: string }
                      required: { type: boolean, default: false }
      responses:
        "200": { $ref: "#/components/responses/SendOK" }
  /forms/{token}:
    get:
      tags: [Interativas]
      summary: "Renderiza a página HTML do formulário (rota pública, sem API key)"
      description: "Aberta pela webview no celular do cliente. O token assinado (HMAC) carrega a definição do form."
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: "HTML do formulário" }
        "400": { description: "Token inválido ou expirado" }
  /forms/{token}/submit:
    post:
      tags: [Interativas]
      summary: "Recebe o preenchimento do formulário (rota pública, sem API key)"
      description: >
        Enviado pela página do form (application/x-www-form-urlencoded). Injeta as
        respostas como mensagem recebida do cliente (Chatwoot + webhook `message`)
        e dispara o webhook `form_response`.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: { type: string }
      responses:
        "200": { description: "Página de confirmação (HTML)" }
        "400": { description: "Token inválido" }
        "410": { description: "Sessão indisponível" }

  # ─────────────────────────────────────────────────────────────────
  # Agendamento (send_at)
  # ─────────────────────────────────────────────────────────────────
...